# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Backend untuk aplikasi Creacipe - Platform berbagi resep masakan.",
        "title": "Creacipe API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/editor/dashboard/stats": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mengambil statistik untuk dashboard editor/admin",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Editor"
                ],
                "summary": "Statistik dashboard",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/editor/menus": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Menampilkan semua resep untuk moderasi (semua status)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Editor"
                ],
                "summary": "Semua resep (moderasi)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter: pending, approved, rejected",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/editor/menus/pending": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Menampilkan resep yang menunggu persetujuan",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Editor"
                ],
                "summary": "Resep pending",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/editor/menus/{id}/status": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mengubah status resep (approve/reject) oleh editor/admin",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Editor"
                ],
                "summary": "Approve/reject resep",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID resep",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Status baru",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UpdateMenuStatusInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.UpdateMenuStatusInput": {
            "type": "object",
            "required": [
                "status"
            ],
            "properties": {
                "rejection_reason": {
                    "type": "string"
                },
                "status": {
                    "type": "string",
                    "enum": [
                        "approved",
                        "rejected"
                    ]
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Masukkan token dengan format: Bearer {access_token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}